+++ omnic-bridge/evm/contracts/interfaces/IBridge.sol
@@ -4,22 +4,22 @@ pragma solidity ^0.8.9;
 
 interface IBridge {
     function swap(
-        uint16 _srcChainId,
+        uint32 _srcChainId,
         uint256 _srcPoolId,
-        uint16 _dstChainId,
+        uint32 _dstChainId,
         uint256 _dstPoolId,
         uint256 _amountLD,
         bytes32 _to
     ) external;
 
     function addLiquidity(
-        uint16 _srcChainId,
+        uint32 _srcChainId,
         uint256 _srcPoolId,
         uint256 _amount
     ) external;
 
     function removeLiquidity(
-        uint16 _srcChainId,
+        uint32 _srcChainId,
         uint256 _srcPoolId,
         uint256 _amount
     ) external;
